+++ harvest_transformer/__main__.py
@@ -71,7 +71,7 @@ def get_new_catalog_id_from_target(target: str) -> str:
     """Extract catalog ID from target"""
     # Currently take catalog_id directly under top-level catalog,
     # as current harvested catalogs do not support nesting
-    new_id = target.split("/")[-1]
+    new_id = target.rstrip("/").split("/")[-1]
 
     if new_id == "":
         return None
@@ -386,7 +386,9 @@ if __name__ == "__main__":
     )
 
     producer = client.create_producer(
-        topic=f"transformed{identifier}", producer_name=f"transformer{identifier}"
+        topic=f"transformed{identifier}",
+        producer_name=f"transformer{identifier}",
+        chunking_enabled=True,
     )
 
     main()